refactor: organization of cookie-related actions
