+++ nextjs-app/src/lib/auth/getCurrentUser.tsx
