+++ nextjs-app/src/lib/auth/hasAccessToken.tsx
